Store information about a file in each node
It may seem pretty redundant to put the name of the file
on every node of the AST, but for now i don't have
any data structure around an individual file (or module).
Besides, the time for optimisation will come. Before that
I should have passing unit and integration tests
that are run from core-lib (so according to the new architecture)